deprecate patch-update flag
