+++ requirements.txt
@@ -1,4 +1,6 @@
 
-pyyaml==5.4.1
-GitPython==3.1.27
-requests==2.25.1
+pyyaml==6.0
+GitPython==3.1.29
+
+# currently not in use since patch-update support dropped
+#requests==2.28.1
